fix bug in liveness_theorem
