+++ src/v2/controllers/vdeployment_controller/model/mod.rs
@@ -1,1 +1,2 @@
 pub mod reconciler;
+pub mod install;
