+++ src/v2/controllers/vdeployment_controller/trusted/mod.rs
@@ -5,3 +5,4 @@ pub mod spec_types;
 pub mod rely_guarantee;
 pub mod step;
 pub mod util;
+pub mod liveness_theorem;
